Deleted and combined smaller .js files for tidier filing
