+++ index.html
@@ -997,10 +997,8 @@
     <script type="module" src="js/circleToggle.js"></script>
     <script type="module" src="js/editButtonHandler.js"></script>
     <script type="module" src="js/diceRoller.js"></script>
-    <script type="module" src="js/resizeHandle.js"></script>
     <script type="module" src="js/CharacterSheetCalulations.js"></script>
-    <script src="js/upload_download.js"></script>
-    <script src="js/page-title.js"></script>
+    <script src="js/pageUtils.js"></script>
 
 </body>
 </html>
